szintlepes megvalositasa

+++ source/unit/gameUnit/GameUnit.h
@@ -3,16 +3,17 @@
 #include <SDL3/SDL_pixels.h>
 #include <SDL3/SDL_rect.h>
 
+#include "../../variables/Variables.h"
 
 class GameUnit {
     float price{};
     int tipus{};
-    int hp{};
-    int maxHp{};
+    float hp{};
+    float maxHp{};
     int range{};
-    int dmg{};
+    float dmg{};
     int poz{};
-    int xp{};
+    float xp{};
     bool isEnemy;
     SDL_FRect rect{};
     SDL_Color color{};
@@ -23,24 +24,24 @@ public:
 
     float getPrice();
     int getTipus();
-    int getHp();
-    int getMaxHp();
-    int getDmg();
+    float getHp();
+    float getMaxHp();
+    float getDmg();
     int getRange();
     int getPos();
-    int getXp();
+    float getXp();
     SDL_FRect getRect();
     SDL_Color getColor();
     float getRextX();
     bool getIsEnemy();
 
     void setPrice(float _price);
-    void setHp(int _hp);
-    void setMaxHp(int _maxHp);
-    void setDmg(int _dmg);
+    void setHp(float _hp);
+    void setMaxHp(float _maxHp);
+    void setDmg(float _dmg);
     void setRange(int _range);
     void setPos(int _pos);
-    void setXp(int _xp);
+    void setXp(float _xp);
     void setRect(SDL_FRect _rect);
     void setColor(SDL_Color _color);
     void setRectX(float x);
